+++ src/index.js
@@ -30,13 +30,14 @@ import "assets/css/material-dashboard-react.css?v=1.9.0";
 const hist = createBrowserHistory();
 
 ReactDOM.render(
-  <Router history={hist}>
-    <Switch>
-      <Route path="/admin" component={Admin} />
-      <Route path="/rtl" component={RTL} />
-      <Route path="/login" component={Login} />
-      <Redirect from="/" to="/login" />
-    </Switch>
-  </Router>,
-  document.getElementById("root")
+  
+    <Router history={hist}>
+      <Switch>
+        <Route path="/admin" component={Admin} />
+        <Route path="/rtl" component={RTL} />
+        <Route path="/login" component={Login} />
+        <Redirect from="/" to="/login" />
+      </Switch>
+    </Router>,
+    document.getElementById("root")
 );