+++ src/views/Login/Login.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import { Link, withRouter } from "react-router-dom";
+import { withRouter } from "react-router-dom";
 import Grid from '@material-ui/core/Grid';
 import api from "../../Services/api";
 import { login } from "../../Services/auth";